changes for eliminating two windows
