+++ SillyBattleSimulation/ViewModels/BaseViewModel.cs
@@ -4,6 +4,7 @@
 
 namespace SillyBattleSimulation.ViewModels
 {
+    using SillyBattleSimulation.Models;
     using System;
     using System.Collections.Generic;
     using System.ComponentModel;
@@ -38,6 +39,12 @@ namespace SillyBattleSimulation.ViewModels
             }
 
             return false;
+        }
+
+        public TeamModel Unload(int team)
+        {
+            TeamModel team = new TeamModel();
+            return team;
         }
     }
 }
